+++ src/ts/main.ts
@@ -1,9 +1,11 @@
+import EntityManager from "./engine/entitiy/entitymanager.js";
+import Renderer from "./engine/renderer/renderer.js";
 import Game from "./game.js";
-import EntityManager from "./engine/entitiy/entitymanager";
 
 //Initialize required stuff
 let entityManager: EntityManager = new EntityManager();
-let game: Game = new Game(entityManager);
+let renderer: Renderer = new Renderer();
+let game: Game = new Game(entityManager, renderer);
 
 // Register event to start the game when the page is fully loaded
 window.addEventListener("load", game.startGame);
